+++ message_ix_models/tests/model/transport/test_demand.py
@@ -2,16 +2,14 @@ import logging
 import re
 from pathlib import Path
 
-import genno
-import pytest
 from genno import ComputationError, Key
 from genno.testing import assert_units
-from pytest import param
+from pytest import mark, param
 
+from message_ix_models import Context
 from message_ix_models.model.structure import get_codes
 from message_ix_models.model.transport import Config, build, demand, testing, workflow
 from message_ix_models.model.transport import plot as plots
-from message_ix_models.model.transport.testing import MARK, make_mark
 from message_ix_models.model.workflow import STAGE
 from message_ix_models.project.ssp import SSP_2017, SSP_2024
 from message_ix_models.report import plot
@@ -19,12 +17,12 @@ from message_ix_models.report import plot
 log = logging.getLogger(__name__)
 
 
-pytestmark = MARK[10]
+pytestmark = mark.transport_build_data
 
 
-@pytest.mark.parametrize("regions", ["R11", "R14", "ISR"])
-@pytest.mark.parametrize("years", ["A", "B"])
-def test_demand_dummy(test_context, regions, years):
+@mark.parametrize("regions", ["R11", "R14", "ISR"])
+@mark.parametrize("years", ["A", "B"])
+def test_demand_dummy(test_context: Context, regions: str, years: str) -> None:
     """Consumer-group-specific commodities are generated."""
     ctx = test_context
     ctx.model.regions = regions
@@ -34,7 +32,7 @@ def test_demand_dummy(test_context, regions, years):
 
     spec = config.spec
 
-    args = (
+    args: tuple[list, list, list, dict] = (
         spec.add.set["commodity"],
         spec.require.set["node"],
         get_codes(f"year/{years}"),  # FIXME should be present in the spec
@@ -52,10 +50,10 @@ def test_demand_dummy(test_context, regions, years):
 
 
 @build.get_computer.minimum_version
-@pytest.mark.parametrize(
+@mark.parametrize(
     "ssp",
     [
-        pytest.param(SSP_2017["2"], marks=make_mark[2](genno.ComputationError)),
+        SSP_2017["2"],
         SSP_2024["1"],
         SSP_2024["2"],
         SSP_2024["3"],
@@ -63,7 +61,9 @@ def test_demand_dummy(test_context, regions, years):
         SSP_2024["5"],
     ],
 )
-def test_exo_pdt(test_context, ssp, regions="R12", years="B"):
+def test_exo_pdt(
+    test_context: Context, ssp: SSP_2017 | SSP_2024, regions="R12", years="B"
+) -> None:
     from message_ix_models.model.transport.testing import assert_units
 
     c, info = testing.configure_build(
@@ -101,7 +101,7 @@ def test_exo_pdt(test_context, ssp, regions="R12", years="B"):
 
 
 @build.get_computer.minimum_version
-def test_exo_report(test_context, tmp_path):
+def test_exo_report(test_context: Context, tmp_path: Path) -> None:
     """Exogenous demand results can be plotted.
 
     Separated from the above because the plotting step is slow.
@@ -133,18 +133,26 @@ def test_exo_report(test_context, tmp_path):
 
 
 @build.get_computer.minimum_version
-@pytest.mark.parametrize(
+@mark.parametrize(
     "regions",
     [
-        param("ISR", marks=MARK[3]),
+        param("ISR", marks=mark.ISR_no_data),
         "R11",
         "R12",
-        param("R14", marks=make_mark[2]((AttributeError, ComputationError))),
+        param(
+            "R14", marks=mark.no_data("node=R14", (AttributeError, ComputationError))
+        ),
     ],
 )
-@pytest.mark.parametrize("years", ["B"])
-@pytest.mark.parametrize("pop_scen", [SSP_2017["2"], SSP_2024["2"]])
-def test_cg_shares(test_context, tmp_path, regions, years, pop_scen):
+@mark.parametrize("years", ["B"])
+@mark.parametrize("pop_scen", [SSP_2017["2"], SSP_2024["2"]])
+def test_cg_shares(
+    test_context: Context,
+    tmp_path: Path,
+    regions: str,
+    years: str,
+    pop_scen: SSP_2017 | SSP_2024,
+) -> None:
     options = dict(ssp=pop_scen)
     c, info = testing.configure_build(
         test_context, tmp_path=tmp_path, regions=regions, years=years, options=options
@@ -179,8 +187,8 @@ R11_WEU  2100  300
 
 @build.get_computer.minimum_version
 def test_pdt_per_capita(
-    tmp_path, test_context, regions="R12", years="B", options=dict()
-):
+    tmp_path: Path, test_context: Context, regions="R12", years="B", options=dict()
+) -> None:
     """Test calculation of PDT per capita, as configured by :func:`.pdt_per_capita`.
 
     Moved from :mod:`.test_operator`.
@@ -202,23 +210,25 @@ def test_pdt_per_capita(
 
 
 @build.get_computer.minimum_version
-@pytest.mark.parametrize(
+@mark.parametrize(
     "regions,years,pop_scen",
     [
-        pytest.param("R11", "A", "GEA mix", marks=MARK[4]),
-        pytest.param("R11", "A", "GEA supply", marks=MARK[4]),
-        pytest.param("R11", "A", "GEA eff", marks=MARK[4]),
+        param("R11", "A", "GEA mix", marks=mark.GEA_not_implemented),
+        param("R11", "A", "GEA supply", marks=mark.GEA_not_implemented),
+        param("R11", "A", "GEA eff", marks=mark.GEA_not_implemented),
         # Different years
-        pytest.param("R11", "B", "GEA mix", marks=MARK[4]),
+        param("R11", "B", "GEA mix", marks=mark.GEA_not_implemented),
         # Different regions & years
         ("R12", "B", SSP_2024["2"]),
         ("R14", "B", SSP_2017["1"]),
         ("R14", "B", SSP_2017["2"]),
         ("R14", "B", SSP_2017["3"]),
-        pytest.param("ISR", "B", SSP_2024["2"], marks=MARK[3]),
+        param("ISR", "B", SSP_2024["2"], marks=mark.ISR_no_data),
     ],
 )
-def test_urban_rural_shares(test_context, tmp_path, regions, years, pop_scen):
+def test_urban_rural_shares(
+    test_context: Context, tmp_path: Path, regions: str, years: str, pop_scen: str
+) -> None:
     options = dict(ssp=pop_scen)
     c, info = testing.configure_build(
         test_context, tmp_path=tmp_path, regions=regions, years=years, options=options
@@ -234,22 +244,24 @@ def test_urban_rural_shares(test_context, tmp_path, regions, years, pop_scen):
     assert set(["UR+SU", "RU"]) == set(result.coords["area_type"].values)
 
 
-@make_mark["gh"](471)
-@MARK["#375"]
+@mark.gh_471
+@mark.gh_375
 @build.get_computer.minimum_version
 @workflow.generate.minimum_version
-@pytest.mark.parametrize(
+@mark.parametrize(
     "nodes, target",
     [
-        pytest.param("R11", "GEA mix", marks=MARK[4]),
+        param("R11", "GEA mix", marks=mark.GEA_not_implemented),
         ("R12", "SSP2"),
         ("R12", "SSP5"),
         ("R14", "SSP2"),
-        pytest.param("R14", "SSP5", marks=make_mark[2](RuntimeError)),
-        pytest.param("R11", "SHAPE innovation", marks=MARK[4]),
+        param("R14", "SSP5", marks=mark.no_data("node=R14, ssp=SSP5", RuntimeError)),
+        param("R11", "SHAPE innovation", marks=mark.SHAPE_not_implemented),
     ],
 )
-def test_cli(tmp_path, mix_models_cli, test_context, nodes, target):
+def test_cli(
+    tmp_path: Path, mix_models_cli, test_context: Context, nodes: str, target: str
+) -> None:
     """Transport CLI can be used to generate build-phase debug outputs."""
     # NB test_context is necessary so that the temporary, in-memory platform established
     #    by .transport.workflow.generate() does not carry to other tests
@@ -258,7 +270,9 @@ def test_cli(tmp_path, mix_models_cli, test_context, nodes, target):
 
     # Identify the path containing the outputs
     expr = re.compile(r"Save to (.*)\.pdf$", flags=re.MULTILINE)
-    output_dir = Path(expr.search(result.output).group(1)).parent
+    match = expr.search(result.output)
+    assert match
+    output_dir = Path(match.group(1)).parent
 
     # Files created in the temporary path
     assert 6 <= len(list(output_dir.glob("*.csv")))